🐛 - fix cannot create movies, and update and create watchlists

+++ app/Http/Controllers/Api/V1/MovieController.php
@@ -7,30 +7,85 @@ use App\Http\Requests\V1\StoreMovieRequest;
 use App\Http\Resources\V1\MovieCollection;
 use App\Http\Resources\V1\MovieResource;
 use App\Models\Movie;
+use App\Traits\HttpResponses;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Arr;
 
 class MovieController extends Controller
 {
+    use HttpResponses;
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $movies = Movie::with('watchlist')->get();
-        return new MovieCollection($movies);
+        try {
+            $movies = Movie::whereHas('watchlists', function ($query) {
+                $query->where('user_id', Auth::id());
+            })->get();
+
+            return $this->success(
+                MovieResource::collection($movies),
+                'Movies retrieved successfully.'
+            );
+        } catch (\Exception $e) {
+            return $this->error(
+                $e->getMessage(),
+                "Failed to user's movies.",
+                500
+            );
+        }
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreMovieRequest $request) {}
+    public function store(StoreMovieRequest $request)
+    {
+        try {
+            $validated = $request->validated();
+            Movie::insert($validated);
+
+            return $this->success(
+                $validated,
+                'Movies added successfully.',
+                201
+            );
+        } catch (\Exception $e) {
+            return $this->error(
+                'Failed to add movies.',
+                $e->getMessage(),
+                500
+            );
+        }
+    }
 
     /**
      * Display the specified resource.
      */
     public function show(Movie $movie)
     {
-        return new MovieResource($movie->load('watchlist'));
+        try {
+            if ($movie->watchlists->user_id !== Auth::id()) {
+                return $this->error(
+                    null,
+                    'Unauthorized access to movie.',
+                    403
+                );
+            }
+
+            return $this->success(
+                new MovieResource($movie),
+                'Movie retrieved successfully.'
+            );
+        } catch (\Exception $e) {
+            return $this->error(
+                null,
+                'Failed to retrieve movie.',
+                500
+            );
+        }
     }
 
     /**
@@ -38,7 +93,27 @@ class MovieController extends Controller
      */
     public function destroy(Movie $movie)
     {
-        $movie->delete();
-        return response()->json(['message' => 'Movie removed successfully.']);
+        try {
+            if ($movie->watchlists->user_id !== Auth::id()) {
+                return $this->error(
+                    null,
+                    'Unauthorized access to delete movie.',
+                    403
+                );
+            }
+
+            $movie->delete();
+
+            return $this->success(
+                null,
+                'Movie deleted successfully.'
+            );
+        } catch (\Exception $e) {
+            return $this->error(
+                $e->getMessage(),
+                'Failed to delete movie.',
+                500
+            );
+        }
     }
 }
